op-supervisor: Move log db to subpackage of db.
