+++ op-supervisor/supervisor/backend/db/logs/db_test.go
@@ -1,4 +1,4 @@
-package db
+package logs
 
 import (
 	"bytes"
