+++ op-supervisor/supervisor/backend/db/logs/entries.go
@@ -1,4 +1,4 @@
-package db
+package logs
 
 import (
 	"encoding/binary"
